# webpos-ordering-service/src/main/resources/api.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Pos
  license:
    name: MIT
paths:
  /cart:
    get:
      summary: List all products in the cart
      operationId: getCart
      tags:
        - cart
      responses:
        '200':
          description: A paged array of products and their quantities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartDto"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /cart/{productId}:
    post:
      summary: Add a product to the cart
      operationId: addToCart
      tags:
        - cart
      parameters:
        - name: productId
          in: path
          required: true
          description: The id of the product to add
          schema:
            type: string
        - name: quantity
          in: query
          description: The quantity of the product to add
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartDto"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Remove a product from the cart
      operationId: removeFromCart
      tags:
        - cart
      parameters:
        - name: productId
          in: path
          required: true
          description: The id of the product to remove
          schema:
            type: string
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartDto"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      summary: Update a product's quantity in the cart
      operationId: updateCart
      tags:
        - cart
      parameters:
        - name: productId
          in: path
          required: true
          description: The id of the product to update
          schema:
            type: string
        - name: quantity
          in: query
          required: true
          description: The new quantity of the product
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartDto"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    CartDto:
      type: array
      items:
        $ref: "#/components/schemas/CartItemDto"
    CartItemDto:
      type: object
      required:
        - productId
        - quantity
      properties:
        productId:
          type: string
        quantity:
          type: integer
          format: int32
    ProductDto:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: number
          format: double
        stock:
          type: integer
          format: int32
        quantity:
          type: integer
          format: int32
        image:
          type: string
    Products:
      type: array
      items:
        $ref: "#/components/schemas/ProductDto"
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string